fix(network): MacOS gets different types in std::min for std::uint64_t and std::size_t
std::uint64_t and std::size_t end up being `unsigned long` and
`unsigned long long` on MacOS, and despite the two types being of the
same size, they are not compatible when invoking templates, so we have
to make sure both arguments are of the same type when invoking std::min

+++ src/reliable-udp/ReliableUdpConnection.cpp
@@ -321,8 +321,9 @@ void ReliableUdpConnection::receiveThreadFunction()
                     individualConnection->second.bufferedBytes.erase(
                         individualConnection->second.bufferedBytes.begin(),
                         individualConnection->second.bufferedBytes.begin() +
-                            std::min(ReliableUdpConnection::packetHeaderSize + receivedDataSize,
-                                     individualConnection->second.bufferedBytes.size()));
+                            std::min(
+                                static_cast<std::size_t>(ReliableUdpConnection::packetHeaderSize + receivedDataSize),
+                                individualConnection->second.bufferedBytes.size()));
                 }
             }
             this->individualConnectionsCondVar.notify_all();
